requirements.txt
flask
python-dotenv
pinecone-client
langchain
openai
werkzeug